requirements.txt
Django==5.2.4
anthropic==0.84.0
python-dotenv==1.1.1
